requirements.txt
apturl==0.5.2
atari-py==0.0.19
beautifulsoup4==4.4.1
blinker==1.3
box2d-py==2.3.1
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
feedparser==5.1.3
guacamole==0.9.2
-e git+https://github.com/openai/gym.git@255a261e5d9a048c9ec768ddd6317c70ab2ad09f#egg=gym
html5lib==0.999
httplib2==0.9.1
idna==2.0
imageio==2.1.2
Jinja2==2.8
Keras==2.0.2
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
mujoco-py==0.5.7
numpy==1.12.1
oauthlib==1.0.3
olefile==0.44
onboard==1.2.0
pachi-py==0.0.21
padme==1.1.1
pexpect==4.0.1
Pillow==4.0.0
plainbox==0.25
protobuf==3.1.0.post1
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pyglet==1.2.4
pygobject==3.20.0
PyJWT==1.3.0
PyOpenGL==3.1.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
pyxdg==0.25
PyYAML==3.12
reportlab==3.3.0
requests==2.13.0
scipy==0.19.0
sessioninstaller==0.0.0
six==1.10.0
system-service==0.3
tensorflow-gpu==0.12.1
Theano==0.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.1.0
xdiagnose==3.8.4
xkit==0.0.0
XlsxWriter==0.7.3
